add screenshot, update readme and rename files
